fix: Address code review feedback for application layer
- GetSessionsByProfileQueryHandler: Make StartTime nullable in DTO to preserve
  semantic difference between "never started" and actual timestamps
- ScenarioSpecifications: Use a.Value for Archetype comparison to ensure EF Core
  can translate to SQL server-side
- UserBadge.TriggerValue: Change from int? to float? to preserve decimal precision
  from scoring calculations; update all usages in AwardBadgeUseCase,
  AwardBadgeCommandHandler, and BadgeAwardingService
- UpdateMediaMetadataUseCase: Add XML documentation and implement optimistic
  concurrency control using Version property
- CreateGameSessionUseCase: Use most restrictive age group (Early Childhood)
  as fallback for unparseable age groups to ensure child-safe behavior

+++ packages/domain/Mystira.Domain/Models/Badge.cs
@@ -318,8 +318,9 @@ public class UserBadge : Entity
 
     /// <summary>
     /// Gets or sets the trigger value that earned this badge.
+    /// Stored as float to preserve decimal precision from scoring calculations.
     /// </summary>
-    public int? TriggerValue { get; set; }
+    public float? TriggerValue { get; set; }
 
     /// <summary>
     /// Gets or sets the threshold value for this badge.
